+++ src/decorators.ts
@@ -1,24 +1,62 @@
 import type {ComponentOptions, ComponentType} from './component';
 import type {ScheduleBuilder, SystemGroup} from './schedule';
 import type {SystemType} from './system';
-import type {Type} from './type';
+import {Type} from './type';
 
 interface PropOptions<JSType> {
   type: Type<JSType> | (() => Type<any>);
   default?: JSType;
 }
 
+function addFieldSchema(options: PropOptions<unknown>, target: any, name: string): void {
+  if (!target.constructor.schema) target.constructor.schema = {};
+  target.constructor.schema[name] = options;
+}
 
-export function field<JSType>(
-  practicalOptions: PropOptions<JSType> | Type<any> | (() => Type<any>)
+
+export function field(
+  practicalOptions: PropOptions<unknown> | Type<any> | (() => Type<any>)
 ) {
   return function(target: any, name: string): void {
-    if (!target.constructor.schema) target.constructor.schema = {};
-    const options: PropOptions<JSType> =
+    const options: PropOptions<unknown> =
       'type' in practicalOptions ? practicalOptions : {type: practicalOptions};
-    target.constructor.schema[name] = options;
+    addFieldSchema(options, target, name);
   };
 }
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+type NotFunction<T> = T extends Function ? never : T;
+function backrefs(type?: ComponentType<any>, fieldName?: string, trackDeletedBackrefs?: boolean):
+  (target: any, name: string) => void;
+function backrefs<T>(target: NotFunction<T>, name: string): void;
+function backrefs(
+  ...args: any[]
+): ((target: any, name: string) => void) | void {
+  if (typeof args[0] === 'function' || args[0] === undefined) {
+    return addFieldSchema.bind(null, {type: Type.backrefs(...args)});
+  }
+  addFieldSchema({type: Type.backrefs}, args[0], args[1]!);
+}
+
+field.boolean = addFieldSchema.bind(null, {type: Type.boolean});
+field.uint8 = addFieldSchema.bind(null, {type: Type.uint8});
+field.int8 = addFieldSchema.bind(null, {type: Type.int8});
+field.uint16 = addFieldSchema.bind(null, {type: Type.uint16});
+field.int16 = addFieldSchema.bind(null, {type: Type.int16});
+field.uint32 = addFieldSchema.bind(null, {type: Type.uint32});
+field.int32 = addFieldSchema.bind(null, {type: Type.int32});
+field.float32 = addFieldSchema.bind(null, {type: Type.float32});
+field.float64 = addFieldSchema.bind(null, {type: Type.float64});
+field.staticString = function(choices: string[]) {
+  return addFieldSchema.bind(null, {type: Type.staticString(choices)});
+};
+field.dynamicString = function(maxUtf8Length: number) {
+  return addFieldSchema.bind(null, {type: Type.dynamicString(maxUtf8Length)});
+};
+field.ref = addFieldSchema.bind(null, {type: Type.ref});
+field.backrefs = backrefs;
+field.object = addFieldSchema.bind(null, {type: Type.object});
+field.weakObject = addFieldSchema.bind(null, {type: Type.weakObject});
 
 
 export const componentTypes: ComponentType<any>[] = [];